Move constants file at the top level
